% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPreserveSearch.R
\name{addPreserveSearch}
\alias{addPreserveSearch}
\title{addPreserveSearch}
\usage{
addPreserveSearch(x)
}
\arguments{
\item{x}{selectizeInput to add preserve search to}
}
\description{
Permite a selectizeInput mantener la busqueda
}
