% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conexion.R
\name{conexion}
\alias{conexion}
\title{R6 Class encapsula las necesidades basicas de una conexion a Postgres}
\description{
R6 Class encapsula las necesidades basicas de una conexion a Postgres

R6 Class encapsula las necesidades basicas de una conexion a Postgres
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{conn}}{Conexión DBI para pasar a funciones}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{conexion$new()}}
\item \href{#method-execute}{\code{conexion$execute()}}
\item \href{#method-send_query}{\code{conexion$send_query()}}
\item \href{#method-list_tables}{\code{conexion$list_tables()}}
\item \href{#method-write_table}{\code{conexion$write_table()}}
\item \href{#method-remove_table}{\code{conexion$remove_table()}}
\item \href{#method-escape}{\code{conexion$escape()}}
\item \href{#method-tbl}{\code{conexion$tbl()}}
\item \href{#method-clone}{\code{conexion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Crea nuevo objeto de conexión
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{conexion$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Esta función requiere unas variables de ambiente cargadas para poder
functionar. DATABASE_NAME para el nombre de la base de datos;
DATABASE_USER para el usuario de autenticación;
DATABASE_PW para la contraseña de dicho usuario;
DATABASE_HOST para la dirección IP de la base de datos;
DATABASE_PORT para el puerto de la base de datos.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\if{latex}{\out{\hypertarget{method-execute}{}}}
\subsection{Method \code{execute()}}{
Método para DBI::dbExecute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{conexion$execute(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Argumentos de DBI::dbExecute}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send_query"></a>}}
\if{latex}{\out{\hypertarget{method-send_query}{}}}
\subsection{Method \code{send_query()}}{
Método para DBI::dbGetQuery
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{conexion$send_query(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Argumentos de DBI::dbGetQuery}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_tables"></a>}}
\if{latex}{\out{\hypertarget{method-list_tables}{}}}
\subsection{Method \code{list_tables()}}{
Método para DBI::dbListTables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{conexion$list_tables()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Argumentos de DBI::dbListTables}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_table"></a>}}
\if{latex}{\out{\hypertarget{method-write_table}{}}}
\subsection{Method \code{write_table()}}{
Método para DBI::dbWriteTable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{conexion$write_table(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Argumentos de DBI::dbWriteTable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_table"></a>}}
\if{latex}{\out{\hypertarget{method-remove_table}{}}}
\subsection{Method \code{remove_table()}}{
Método para DBI::dbRemoveTable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{conexion$remove_table(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Argumentos de DBI::dbRemoveTable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-escape"></a>}}
\if{latex}{\out{\hypertarget{method-escape}{}}}
\subsection{Method \code{escape()}}{
Método para DBI::sqlInterpolate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{conexion$escape(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Argumentos de DBI::sqlInterpolate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tbl"></a>}}
\if{latex}{\out{\hypertarget{method-tbl}{}}}
\subsection{Method \code{tbl()}}{
Método para dplyr::tbl
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{conexion$tbl(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Argumentos de dplyr::tbl}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{conexion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
