% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtros.R
\name{filtros_discretos_server}
\alias{filtros_discretos_server}
\title{Servidor para filtros discretos}
\usage{
filtros_discretos_server(id, datos, cache, max_char = 20)
}
\arguments{
\item{id}{ID del módulo}

\item{datos}{Objecto de class "tabla_mutable"}

\item{cache}{ReactiveValues para el cache de shinyCache}

\item{max_char}{Número de filtros máximo}
}
\description{
Filtros se aplican con un trigger con el ID del módulo.
}
