% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtros.R
\name{filtros_discretos_ui}
\alias{filtros_discretos_ui}
\title{UI para filtros discretos}
\usage{
filtros_discretos_ui(id)
}
\arguments{
\item{id}{ID del módulo}
}
\value{
Objeto tags para UI de una aplicación Shiny
}
\description{
UI para filtros discretos
}
