% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtros-numericos.R
\name{filtros_numericos_server}
\alias{filtros_numericos_server}
\title{Servidor para filtros numéricos}
\usage{
filtros_numericos_server(id, datos, max_num = 20)
}
\arguments{
\item{id}{ID del módulo}

\item{datos}{Objecto de class "tabla mutable"}

\item{max_num}{Número de filtros máximo}
}
\description{
Filtros se aplican con un trigger con el ID del módulo.
}
