% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtros-numericos.R
\name{filtros_numericos_ui}
\alias{filtros_numericos_ui}
\title{UI para filtros numéricos}
\usage{
filtros_numericos_ui(id)
}
\arguments{
\item{id}{ID del módulo}
}
\value{
Objeto tags para UI de una aplicación Shiny
}
\description{
UI para filtros numéricos
}
