% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla.R
\name{tabla}
\alias{tabla}
\title{R6 Class encapsula los datos y métodos para una tabla mutable}
\description{
Un objeto creado con esta clase podrá ser mutado de manera sencilla
agregando modificaciones y aplicandolas. Modificaciones especificas
se pueden agregar, modificar o eliminar con los métodos encapsulados aqui.
Las modificaciones se pasan con una función de un solo argumento que
modifique los datos. Cada modificación lleva una prioridad, entre más alta
esta prioridad, se aplicará antes.
}
\examples{
example_data <- data.frame(n = 1:100)
number_table <- tabla$new(example_data)
number_table$mod_add(
  id = "x2",
  priority = 1,
  .f = function(.data) {
    .data$n <- .data$n * 2
    return(.data)
  }
)
number_table$mod_aplicar()
print(number_table$table)
number_table$reset()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mod_list}}{Lista de modificaciones para hacer a la tabla}

\item{\code{base}}{Los datos base (originales) del objeto.}

\item{\code{tabla}}{La tabla en estado mutable.}

\item{\code{colnames}}{Las columnas de los datos base}

\item{\code{colnames_num}}{Las columnas de tipo númerico de los datos base}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tabla$new()}}
\item \href{#method-reset}{\code{tabla$reset()}}
\item \href{#method-mod_reset}{\code{tabla$mod_reset()}}
\item \href{#method-mod_add}{\code{tabla$mod_add()}}
\item \href{#method-mod_rm}{\code{tabla$mod_rm()}}
\item \href{#method-mod_aplicar}{\code{tabla$mod_aplicar()}}
\item \href{#method-query}{\code{tabla$query()}}
\item \href{#method-clone}{\code{tabla$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Crea una tabla mutable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tabla$new(.data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{Tibble, Data.table, Tbl SQL, etc...}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Devuelve la tabla a su estado original
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tabla$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mod_reset"></a>}}
\if{latex}{\out{\hypertarget{method-mod_reset}{}}}
\subsection{Method \code{mod_reset()}}{
Elimina todas las modificaciones a la tabla
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tabla$mod_reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mod_add"></a>}}
\if{latex}{\out{\hypertarget{method-mod_add}{}}}
\subsection{Method \code{mod_add()}}{
Agrega una modificación para hacerse a la tabla
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tabla$mod_add(id, priority, .f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Un ID único para la modificación}

\item{\code{priority}}{Nivel de prioridad de la modificación (define el orden)}

\item{\code{.f}}{función con un solo argumento para aplicar a la tabla}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mod_rm"></a>}}
\if{latex}{\out{\hypertarget{method-mod_rm}{}}}
\subsection{Method \code{mod_rm()}}{
Remueve alguna modificación de la lista
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tabla$mod_rm(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{El ID único de la modificación}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mod_aplicar"></a>}}
\if{latex}{\out{\hypertarget{method-mod_aplicar}{}}}
\subsection{Method \code{mod_aplicar()}}{
Aplica todas las modificaciones en la lista en orden según la prioridad.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tabla$mod_aplicar()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query"></a>}}
\if{latex}{\out{\hypertarget{method-query}{}}}
\subsection{Method \code{query()}}{
Extrae el SQL query o un hash que identifique el estado actual
de la tabla.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tabla$query()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tabla$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
