% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtros-numericos.R
\name{filtros_numericos_server}
\alias{filtros_numericos_server}
\title{Servidor para filtros numéricos}
\usage{
filtros_numericos_server(id, datos, max_num = 20)
}
\arguments{
\item{id}{ID del módulo}

\item{datos}{Objecto de class "tabla_mutable"}

\item{max_num}{Número de filtros máximo}
}
\description{
Filtros se aplican con un trigger con el ID del módulo.
}
\details{
Para poder aplicar los filtros debe llamar a un trigger con el ID del módulo.
En caso de que se actualice el objeto de tabla mutable, se trendra que llamar
un trigger con el siguiente formato id-actualizar. Este reseteará el
UI del módulo para adaptarse a la nueva tabla.
}
\examples{
# Para actualizar el módulo a una nueva tabla mutable
gargoyle::trigger("id-actualizar")
# Para aplicar los filtros y meterlos en la lista de modificaciones
gargoyle::trigger("id")
}
